Inline unused super class for server preferences
